backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.24.3
scipy==1.11.4
opencv-python-headless==4.8.1.78
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
fingerprint-enhancer==0.0.14
